feat: teacher-main.tsx 파일을 여러 컴포넌트로 리팩토링
